correct maven setup inshallah
