--- Swiftify/src/main/java/org/switify/App.java
@@ -1,13 +1,0 @@
-package org.switify;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
