+++ swiftify-backend/src/main/java/com/group113/swiftify/DependencyFactory.java
@@ -1,0 +1,26 @@
+
+package com.group113.swiftify;
+
+import software.amazon.awssdk.auth.credentials.EnvironmentVariableCredentialsProvider;
+import software.amazon.awssdk.http.urlconnection.UrlConnectionHttpClient;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.S3Client;
+
+/**
+ * The module containing all dependencies required by the {@link App}.
+ */
+public class DependencyFactory {
+
+    private DependencyFactory() {}
+
+    /**
+     * @return an instance of S3Client
+     */
+    public static S3Client s3Client() {
+        return S3Client.builder()
+                       .credentialsProvider(EnvironmentVariableCredentialsProvider.create())
+                       .region(Region.US_EAST_1)
+                       .httpClientBuilder(UrlConnectionHttpClient.builder())
+                       .build();
+    }
+}
