+++ swiftify-backend/src/test/java/com/group113/swiftify/AppTest.java
@@ -1,0 +1,15 @@
+package com.group113.swiftify;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
+
+public class AppTest {
+
+    @Test
+    public void handleRequest_shouldReturnConstantValue() {
+        App function = new App();
+        Object result = function.handleRequest("echo", null);
+        assertEquals("echo", result);
+    }
+}
