+++ swiftify-backend/template.yaml
@@ -1,0 +1,19 @@
+AWSTemplateFormatVersion: '2010-09-09'
+Transform: AWS::Serverless-2016-10-31
+Resources:
+  # See https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/sam-resource-function.html
+  # for more info to see how to tune the lambda function configs based on your use case.
+  AppFunction:
+    Type: AWS::Serverless::Function
+    Properties:
+      Runtime: java17
+      Handler: com.group113.swiftify.App::handleRequest
+      Timeout: 60
+      MemorySize: 512
+      CodeUri: ./target/swiftify-backend.jar
+      # Attach policies here to give the function permission to access other AWS resources if needed
+      # See: https://github.com/awslabs/serverless-application-model/blob/master/docs/policy_templates.rst
+      # eg:
+      #Policies:
+      # - S3ReadPolicy:
+      #     BucketName: test-bucket
