+++ swiftify-frontend/src/main/java/com/group113/swiftify/App.java
@@ -1,0 +1,13 @@
+package com.group113.swiftify;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
